研二上学期开始时的进展

+++ src/util/DTWDistance.java
@@ -2,12 +2,12 @@ package util;
 
 public class DTWDistance{ 
 	private Long[] test;
-	private Long[] signature;
+	private Double[] signature;
 	
-	private Long[][] distanceMatrix;
-	private Long[][] costMatrix;
+	private Double[][] distanceMatrix;
+	private Double[][] costMatrix;
 	
-	public DTWDistance(Long[] signature, Long[] test) {
+	public DTWDistance(Double[] signature, Long[] test) {
 //		this.test = test;
 //		this.signature = signature;
 //		//行对应signatu， 列对应test
@@ -20,12 +20,12 @@ public class DTWDistance{
 	public DTWDistance() {
 	}
 
-	public void setVariable(Long[] signature, Long[] test) {
+	public void setVariable(Double[] signature, Long[] test) {
 		this.test = test;
 		this.signature = signature;
 		//行对应signatu， 列对应test
-		distanceMatrix = new Long[signature.length][test.length];
-		costMatrix = new Long[signature.length][test.length];
+		distanceMatrix = new Double[signature.length][test.length];
+		costMatrix = new Double[signature.length][test.length];
 	
 	}
 	
@@ -38,7 +38,7 @@ public class DTWDistance{
 		return;
 	}
 	
-	public Long getDTWDistance(Long[] sig, Long[] test) {
+	public Double getDTWDistance(Double[] sig, Long[] test) {
 		setVariable(sig, test);
 		constructDistanceMatrix();
 		getCostMatrix();
@@ -49,16 +49,16 @@ public class DTWDistance{
 	/*
 	 * 使用欧式距离（两点间直线距离）
 	 */
-	private Long getDistance(Long a, Long b) {
+	private Double getDistance(Double a, Long b) {
 		return getAbsDiff(a, b);
 	}
 	
-	public Long getAbsDiff(Long a, Long b) {
+	public Double getAbsDiff(Double a, Long b) {
 		if (a > b) return a - b;
 		else return b - a;
 	}
 	
-	public Long getMin(Long a, Long b, Long c) {
+	public Double getMin(Double a, Double b, Double c) {
 		if (a > b) return (b > c) ? c : b;
 		else return (a > c) ? c : a;
 	}
@@ -83,7 +83,7 @@ public class DTWDistance{
 	
 	public static void main(String[] agrs) {
 		Long[] aLong = new Long[2];
-		Long[] signature = {3l, 5l, 6l, 7l,7l, 1l};
+		Double[] signature = {3.0, 5.0, 6.0, 7.0,7.0, 1.0};
 		Long[] test1 = {3l,6l,6l,7l,8l,1l,1l };
 		Long[] test2 = {2l,5l,7l,7l,7l,7l,2l};
